+++ crates/wdl-analysis/src/document/v1.rs
@@ -757,7 +757,7 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
         scopes: vec![Scope::new(
             None,
             definition
-                .braced_scope_span()
+                .braced_scope_span(false)
                 .expect("should have brace scope span"),
         )],
         inputs,
@@ -861,7 +861,7 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
                         definition
                             .output()
                             .expect("should have output section")
-                            .braced_scope_span()
+                            .braced_scope_span(false)
                             .expect("should have braced scope span"),
                         HiddenType::TaskPostEvaluation,
                     )
@@ -877,9 +877,9 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
             TaskGraphNode::Command(section) => {
                 let scope_index = *command_scope.get_or_insert_with(|| {
                     let span = if section.is_heredoc() {
-                        section.heredoc_scope_span()
+                        section.heredoc_scope_span(false)
                     } else {
-                        section.braced_scope_span()
+                        section.braced_scope_span(false)
                     };
 
                     create_section_scope(
@@ -910,7 +910,7 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
                         &mut task.scopes,
                         &name,
                         section
-                            .braced_scope_span()
+                            .braced_scope_span(false)
                             .expect("should have braced scope span"),
                         HiddenType::TaskPreEvaluation,
                     )
@@ -934,7 +934,7 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
                         &mut task.scopes,
                         &name,
                         section
-                            .braced_scope_span()
+                            .braced_scope_span(false)
                             .expect("should have braced scope span"),
                         HiddenType::TaskPreEvaluation,
                     )
@@ -958,7 +958,7 @@ fn add_task(config: &Config, document: &mut DocumentData, definition: &TaskDefin
                         &mut task.scopes,
                         &name,
                         section
-                            .braced_scope_span()
+                            .braced_scope_span(false)
                             .expect("should have braced scope span"),
                         HiddenType::TaskPreEvaluation,
                     )
@@ -1078,7 +1078,7 @@ fn populate_workflow(config: &Config, document: &mut DocumentData, workflow: &Wo
     let mut scopes = vec![Scope::new(
         None,
         workflow
-            .braced_scope_span()
+            .braced_scope_span(false)
             .expect("should have braced scope span"),
     )];
     let mut output_scope = None;
@@ -1160,7 +1160,7 @@ fn populate_workflow(config: &Config, document: &mut DocumentData, workflow: &Wo
                             workflow
                                 .output()
                                 .expect("should have output section")
-                                .braced_scope_span()
+                                .braced_scope_span(false)
                                 .expect("should have braced scope span"),
                         ),
                     )
@@ -1366,7 +1366,7 @@ fn add_conditional_statement(
             Scope::new(
                 Some(parent),
                 clause
-                    .braced_scope_span()
+                    .braced_scope_span(false)
                     .expect("should have braced scope span"),
             ),
         );
@@ -1402,7 +1402,7 @@ fn add_scatter_statement(
         Scope::new(
             Some(parent),
             statement
-                .braced_scope_span()
+                .braced_scope_span(false)
                 .expect("should have braced scope span"),
         ),
     );