+++ crates/wdl-lsp/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added support for the `textDocument/prepareCallHierarchy`, `callHierarchy/incomingCalls`, and
   `callHierarchy/outgoingCalls` requests ([#874](https://github.com/stjude-rust-labs/sprocket/pull/874)).
+* Added support for the `textDocument/foldingRange` request ([#867](https://github.com/stjude-rust-labs/sprocket/pull/867)).
 
 ### Changed